% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optcut_prop.R
\name{optcutoff_prop}
\alias{optcutoff_prop}
\title{A function to calculate the optimal clinical cut-off}
\usage{
optcutoff_prop(npat, cf, scenario, par1 = c(1.33, 0.77, 0.4),
  medianimp = 0.249, cureimp = 0.05, cencut = 8, tpred = 0.25,
  tadd = 0.25, nsamp = 100, fCI = 0, lpi = 2.5, entry = rep(lpi/npat *
  1:npat, 2), al = 0.05)
}
\arguments{
\item{npat}{The total number of patients in the study.}

\item{cf}{A vector representing all the time points (in years) after the last patient in, when the study can be stopped.}

\item{scenario}{A dataframe representing different effect-size scenarios for diff combination of median and cure rate improvement. First column called "impmed" has different values for median improvement and the second column "impcure" represents the different values of cure-rate improvement.}

\item{par1}{The true parametric value for the control group: (k,l,r) where k and l are the shape and scale parameters for the weibull distribution and r is the cure rate for the control treatment. Default values: (1.33,0.77,0.40).}

\item{medianimp}{True improvement in median survival rate for uncured patients while using the experimental treatment. Default value: 0.249.}

\item{cureimp}{True improvement in cure rate while using the experimental treatment. Default value: 0.05.}

\item{cencut}{Time point (in years from the start of the study) after which a patient will no longer be followed.}

\item{tpred}{Time of additional follow up (in years) after study ends. Default value: 0.25.}

\item{tadd}{Time of additional follow up (in years) after the first additional follow up time  in study. Default value: 0.25.}

\item{nsamp}{Number of resamples needed to calculate confidence interval. Default: 100}

\item{fCI}{A dummy to calculate confidence intervals. 1=yes, 0=no. Default: 0}

\item{lpi}{Last patient in (in years calculated from the start of the study). Default: 2.5.}

\item{entry}{Entry time of parients, twice for control group and then treatment group. Default: rep(lpi/npat * 1:npat, 2)}

\item{al}{(1-al)*100 \% CI is calculated. Default value: 0.05.}
}
\value{
It returns a dataframe.

cutoff: The position of the observation in the vector "cf" which is also the estimated cut-off point.

param_k: The estimated parametric value of the shape parameter for the weibull distribution for both treatment groups.

ctrparam_l: The estimated parametric value of the scale parameter for the weibull distribution for control group.

ctrparam_r: The estimated parametric value of the cure rate of the control treatment.

trtparam_l: The estimated parametric value of the scale parameter for the weibull distribution for treatment group.

trtparam_r: The estimated parametric value of the cure rate of the experimental treatment.

ctrhaztp: Total hazard of the risk set of the control group at a future time point.

ctrevetp: The estimated number of events to occur in the control group at additional follow-up time point "tpred" after the end of the study.

ctrevetpa: The estimated number of events to occur in the control group at additional follow-up time point "tpred + tadd" after the end of the study.

ctrevetmax: The estimated maximum number of events to occur in the control group with additional follow-up after the end of the study.

mean_ctrhaz: Mean total hazard of the risk set of the control group at a future time point for the resampled data.

med_ctrhaz: Median total hazard of the risk set of the control group at a future time point for the resampled data.

mean_ctrevep: Mean estimated number of events to occur in the control group at additional follow-up time point "tpred" after the end of the study for the resampled data.

med_ctrevep: Median estimated number of events to occur in the control group at additional follow-up time point "tpred" after the end of the study for the resampled data.

mean_ctrevepa: Mean estimated number of events to occur in the control group at additional follow-up time point "tpred + tadd" after the end of the study for the resampled data.

med_ctrevepa: Median estimated number of events to occur in the control group at additional follow-up time point "tpred + tadd" after the end of the study for the resampled data.

mean_ctrevemax: Mean estimated maximum number of events to occur in the control group with additional follow-up after the end of the study for the resampled data.

med_ctrevemax: Median estimated maximum number of events to occur in the control group with additional follow-up after the end of the study for the resampled data.

ctrhaz_l: Lower limit of the total hazard of the risk set of the control group at a future time point for the resampled data.

ctrhaz_u: Upper limit of the total hazard of the risk set of the control group at a future time point for the resampled data.

ctrevep_l: Lower limit of the estimated number of events to occur in the control group at additional follow-up time point "tpred" after the end of the study for the resampled data.

ctrevep_u: Upper limit of the estimated number of events to occur in the control group at additional follow-up time point "tpred" after the end of the study for the resampled data.

ctrevepa_l: Lower limit of the estimated number of events to occur in the control group at additional follow-up time point "tpred + tadd" after the end of the study for the resampled data.

ctrevepa_u: Upper limit of the estimated number of events to occur in the control group at additional follow-up time point "tpred + tadd" after the end of the study for the resampled data.

ctrevemax_l: Lower limit of the estimated maximum number of events to occur in the control group with additional follow-up after the end of the study for the resampled data.

ctrevemax_u: Upper limit of the estimated maximum number of events to occur in the control group with additional follow-up after the end of the study for the resampled data.
}
\description{
This command helps in calculating the cut-off point in a simulation study for cure rate survival data.
}
\examples{
scenarios <- data.frame(impmed = rep(c(0.2,0.3, 0.4), rep(3, 3)),
                      impcure = rep(c(0.1, 0.05, 0), 3))
cuts <- c(1,1.25,1.3,1.35,1.4,1.45,1.5,1.55,1.6,1.7,1.8)
optcutoff_prop(npat=100,scenario=scenarios,cf=cuts) # gives only estimates
optcutoff_prop(npat=100,scenario=scenarios,cf=cuts,fCI=1) # gives estimates along with confidence intervals
}

